Move TokenDetails to :security
