+++ security/src/main/java/net/explorviz/shared/security/TokenDetails.java
@@ -1,4 +1,4 @@
-package net.explorviz.shared.security.model;
+package net.explorviz.shared.security;
 
 import java.time.ZonedDateTime;
 import java.util.List;
