+++ security/src/main/java/net/explorviz/shared/security/TokenParserService.java
@@ -18,7 +18,6 @@ import java.util.Map;
 import javax.validation.constraints.NotNull;
 import javax.ws.rs.ForbiddenException;
 import net.explorviz.shared.config.annotations.Config;
-import net.explorviz.shared.security.model.TokenDetails;
 
 /**
  * This injectable service is used to extract and parse the details of a JSON web token. If used,
