+++ security/src/main/java/net/explorviz/shared/security/filters/AuthenticationFilter.java
@@ -12,7 +12,7 @@ import javax.ws.rs.ext.Provider;
 import net.explorviz.shared.security.AuthenticatedUserDetails;
 import net.explorviz.shared.security.TokenBasedSecurityContext;
 import net.explorviz.shared.security.TokenParserService;
-import net.explorviz.shared.security.model.TokenDetails;
+import net.explorviz.shared.security.TokenDetails;
 
 /**
  * Custom {@link ContainerRequestFilter} that is used for JWT-based authentication. If used in a web
